Move walkingpadfitbit.auth.usecases classes to walkingpadfitbit.auth.
We don't need usecases here: it's not the domain layer.
